API

Public

Loggers

Memento.Logger — Type.

A Logger is responsible for converting msg strings into Records which are then passed to each handler. By default loggers propagate their message to their parent loggers.

Fields:

name: is the name of the logger (required).

handlers: is a collection of Handlers (defaults to empty Dict).

level: the current minimum logging level for the logger to log message to handlers (defaults to "not_set").

levels: a mapping of available logging levels to their relative priority (represented as integer values) (defaults to using Memento._log_levels)

record: the Record type that should be produced by this logger (defaults to DefaultRecord).

propagate: whether or not this logger should propagate its message to its parent (defaults to true).

source

Base.error — Function.

error(::Logger, ::AbstractString) logs the message at the error level and throws an ErrorException with that message error(::Logger, ::Exception) calls error(logger, msg) with the contents of the Exception.

source

Base.info — Function.

info(::Logger, ::AbstractString) logs the message at the info level.

source

Base.log — Method.

log(::Logger, ::AbstractString, ::AbstractString) creates a Dict with the logger name, level, levelnum and message and calls the other log method (which may recursively call itself on parent loggers with the created Dict).

Args:

logger: the logger to log to.

level: the log level as a String

msg: the msg to log as a String

source

Base.log — Method.

log(::Logger, ::Dict{Symbol, Any}) logs logger.record(args) to its handlers if it has the appropriate args[:level] and args[:level] is above the priority of logger.level. If this logger is not the root logger and logger.propagate is true then the parent logger is called.

Args:

logger: the logger to log args to.

args: a dict of msg fields and values that should be passed to logger.record.

source

Base.warn — Function.

warn(::Logger, ::AbstractString) logs the message at the warn level.

source

Base.warn — Method.

warn(::Logger, ::Exception) takes an exception and logs it.

source

Memento.add_handler — Function.

add_handler(::Logger, ::Handler, name) adds a new handler to logger.handlers. If a name is not provided a random one will be generated.

Args:

logger: the logger to use.

handler: the handler to add.

name: a name to identify the handler.

source

Memento.add_level — Method.

add_level(::Logger, ::AbstractString, ::Int) adds a new level::String and priority::Int to the logger.levels

source

Memento.alert — Function.

alert(::Logger, ::AbstractString) logs the message at the alert level and throws an ErrorException with that message alert(::Logger, ::Exception) calls alert(logger, msg) with the contents of the Exception.

source

Memento.basic_config — Method.

basic_config(::AbstractString; ::AbstractString, ::Dict{AbstractString, Int}, colorized::Bool) sets the Memento._log_levels, creates a default root logger with a DefaultHandler that prints to STDOUT.

Args:

level: the minimum logging level to log message to the root logger (required).

fmt: a format string to pass to the DefaultFormatter which describes how to log messages (defaults to Memento.DEFAULT_FMT_STRING)

levels: the default logging levels to use (defaults to Memento._log_levels).

colorized: whether or not the message to STDOUT should be colorized.

Returns the root logger.

source

Memento.critical — Function.

critical(::Logger, ::AbstractString) logs the message at the critical level and throws an ErrorException with that message critical(::Logger, ::Exception) calls critical(logger, msg) with the contents of the Exception.

source

Memento.debug — Function.

debug(::Logger, ::AbstractString) logs the message at the debug level.

source

Memento.emergency — Function.

emergency(::Logger, ::AbstractString) logs the message at the emergency level and throws an ErrorException with that message emergency(::Logger, ::Exception) calls emergency(logger, msg) with the contents of the Exception.

source

Memento.get_handlers — Method.

get_handlers(::Logger) returns logger.handlers

source

Memento.get_logger — Function.

get_logger(::AbstractString) returns the appropriate logger. If the logger or its parents do not exist then they are initialized with no handlers and not set.

Args: the name of the logger (defaults to "root")

Returns the logger.

source

Memento.get_logger — Method.

get_logger(name::Module) converts the Module to a String and calls get_logger(name::String).

Args: name of the logger

Returns the logger.

source

Memento.is_root — Method.

is_root(::Logger) returns true if logger.nameis "root" or ""

source

Memento.is_set — Method.

is_set(:Logger) returns true or false as to whether the logger is set. (ie: logger.level != "not_set")

source

Memento.notice — Function.

notice(::Logger, ::AbstractString) logs the message at the notice level.

source

Memento.remove_handler — Method.

remove_handler(::Logger, name) removes the Handler with the provided name from the logger.handlers.

source

Memento.set_level — Method.

set_level(::Logger, ::AbstractString) changes what level this logger should log at.

source

Memento.set_record — Method.

set_record{R<:Record}(::Logger, ::Type{R}) sets the record type for the logger.

Args:

logger: the logger to set.

rec: A Record type to use for logging messages (ie: DefaultRecord).

source

Handlers

Memento.DefaultHandler — Type.

DefaultHandler{F<Formatter, O<:IO}(io::O, fmt::F, opts::Dict{Symbol, Any}) creates a DefaultHandler with the specified IO type.

Args:

io: the IO type

fmt: the Formatter to use (default to DefaultFormatter())

opts: the optional arguments (defaults to Dict{Symbol, Any}())

source

Memento.DefaultHandler — Type.

DefaultHandler{F<Formatter}(filename::AbstractString, fmt::F, opts::Dict{Symbol, Any}) creates a DefaultHandler with a IO handle to the specified filename.

Args:

filename: the filename of a log file to write to

fmt: the Formatter to use (default to DefaultFormatter())

opts: the optional arguments (defaults to Dict{Symbol, Any}())

source

Memento.DefaultHandler — Type.

The DefaultHandler manages any Formatter, IO and Record.

Fields:

fmt: a Formatter for converting Records to Strings

io: an IO type for printing String to.

opts: a dictionary of optional arguments such as :is_colorized and :colors Ex) Dict{Symbol, Any}( :is_colorized => true, :opts[:colors] => Dict{AbstractString, Symbol}( "debug" => :blue, "info" => :green, ... ) )

source

Memento.Handler — Type.

Handlers manage formatting Records and printing the resulting String to an IO type. All Handler subtypes must implement at least 1 log(::Handler, ::Record) method.

NOTE: Handlers can useful if you need to special case logging behaviour based on the Formatter, IO and/or Record types.

source

Base.log — Method.

log{F<:Formatter, O<:IO}(handler::DefaultHandler{F ,O}, rec::Record) logs all records with any Formatter and IO types.

source

Base.log — Method.

logs{F<:Formatter, O<:Syslog}(handler::DefaultHandler{F, O}, rec::Record) logs all records with any Formatter and a Syslog IO type.

source

Formatters

Memento.DefaultFormatter — Type.

The DefaultFormatter uses a simple format string to build the log message. Fields from the Record to be used should be wrapped curly brackets.

Ex) "[{level} | {name}]: {msg}" will print message of the form [info | root]: my info message. [warn | root]: my warning message. ...

source

Memento.Formatter — Type.

A Formatter must implement a format(::Formatter, ::Record) method which takes a Record and returns a String representation of the log Record.

source

Memento.JsonFormatter — Type.

JsonFormatter uses the JSON pkg to format the Record into a valid JSON string.

source

Memento.format — Method.

format(::DefaultFormatter, ::Record) iteratively replaces entries in the format string with the appropriate fields in the Record

source

Memento.format — Method.

format(::JsonFormatter, ::Record) converts :date, :lookup and :stacktrace to strings and dicts respectively and call JSON.json() on the resulting dictionary.

source

Records

Memento.DefaultRecord — Type.

DefaultRecord wraps a Dict{Symbol, Any} which stores basic logging event information.

Info:

date: timestamp of log event level: log level levelnum: integer value for log level msg: the log message itself name: the name of the source logger pid: the pid of where the log event occured lookup: the top StackFrame stacktrace: a stacktrace

source

Memento.Record — Type.

Records are used to store information about a log event including the msg, date, level, stacktrace, etc. Formatters use Records to format log message strings.

source

IO

Memento.FileRoller — Type.

A FileRoller is responsible for managing a rolling log file.

Fields:

prefix: filename prefix for the log.

folder: directory where the log should be written.

file: the current file IO handle

byteswritten: keeps track of how many bytes have been written to the current file.

max_sz: the maximum number of bytes written to a file before rolling over to another.

source

Memento.FileRoller — Method.

FileRoller(prefix, dir; max_size=DEFAULT_MAX_FILE_SIZE) creates a rolling log file in the specified directory with the given prefix.

source

Memento.FileRoller — Method.

FileRoller(prefix; max_size=DEFAULT_MAX_FILE_SIZE) creates a rolling log file in the current working directory with the specified prefix.

source

Memento.Syslog — Type.

Syslog handle writing message to syslog by shelling out to the logger command.

Fields:

facility: The syslog facility to write to (e.g., :local0, :ft, :daemon, etc) (defaults to :local0)

tag: a tag to use for all message (defaults to "julia")

pid: tags julia's pid to messages (defaults to -1 which doesn't include the pid)

source

Private

Loggers

Base.show — Method.

Base.show(::IO, ::Logger) just prints Logger(logger.name)

source

Memento.get_parent — Method.

get_parent(::AbstractString) takes a string representing the name of a logger and returns its parent. If the logger name has no parent then the root logger is returned. Parent loggers are extracted assuming a naming convention of "foo.bar.baz", where "foo.bar.baz" is the child of "foo.bar" which is the child of "foo"

Args:

name: the name of the logger.

Returns the parent logger.

source

Memento.reset! — Method.

reset! removes all registered loggers and reinitializes the root logger without any handlers.

source

Handlers

Memento.setup_opts — Method.

setup_opts(opts) sets the default :colors if opts[:is_colorized] == true

source

Formatters

Records

Base.getindex — Method.

getindex(::Record, key) returns the item from the inner dict

source

Base.keys — Method.

keys(::Record) returns all keys in the inner dict

source

Memento.getdict — Method.

getdict(::Record) returns the inner dict of the record

source

IO

Base.flush — Method.

flush(::FileRoller) flushes the current open file.

source

Base.flush — Method.

flush(::Syslog) is defined just in case somebody decides to call flush, which is unnecessary.

source

Base.println — Method.

println(::FileRoller, ::AbstractString) writes the string to a file and creates a new file if we've reached the max file size.

source

Base.println — Method.

println(::Syslog, ::AbstractString, ::AbstractString) converts the first AbstractString to a Symbol and call println(::Syslog, ::Symbol, ::AbstractString)

source

Base.println — Method.

println(::Syslog, ::Symbol, ::AbstractString) writes the AbstractString to logger with the Symbol representing the syslog level.

source

Memento.getfile — Method.

getfile(folder::AbstractString, prefix::AbstractString) grabs the next log file.

source

Memento.getsuffix — Method.

getsuffix(::Integer) formats the nth file suffix.

source