Private

Loggers

Base.show — Method.

Base.show(::IO, ::Logger) just prints Logger(logger.name)

source

Memento.get_parent — Method.

get_parent(::AbstractString) takes a string representing the name of a logger and returns its parent. If the logger name has no parent then the root logger is returned. Parent loggers are extracted assuming a naming convention of "foo.bar.baz", where "foo.bar.baz" is the child of "foo.bar" which is the child of "foo"

Args:

name: the name of the logger.

Returns the parent logger.

source

Memento.reset! — Method.

reset! removes all registered loggers and reinitializes the root logger without any handlers.

source

Handlers

Memento.setup_opts — Method.

setup_opts(opts) sets the default :colors if opts[:is_colorized] == true

source

Formatters

Records

Base.getindex — Method.

getindex(::Record, key) returns the item from the inner dict

source

Base.keys — Method.

keys(::Record) returns all keys in the inner dict

source

Memento.getdict — Method.

getdict(::Record) returns the inner dict of the record

source

IO

Base.flush — Method.

flush(::FileRoller) flushes the current open file.

source

Base.flush — Method.

flush(::Syslog) is defined just in case somebody decides to call flush, which is unnecessary.

source

Base.println — Method.

println(::FileRoller, ::AbstractString) writes the string to a file and creates a new file if we've reached the max file size.

source

Base.println — Method.

println(::Syslog, ::AbstractString, ::AbstractString) converts the first AbstractString to a Symbol and call println(::Syslog, ::Symbol, ::AbstractString)

source

Base.println — Method.

println(::Syslog, ::Symbol, ::AbstractString) writes the AbstractString to logger with the Symbol representing the syslog level.

source

Memento.getfile — Method.

getfile(folder::AbstractString, prefix::AbstractString) grabs the next log file.

source

Memento.getsuffix — Method.

getsuffix(::Integer) formats the nth file suffix.

source